+++ torchrua/layout/left.py
@@ -14,7 +14,7 @@ class LeftAlignedSequence(NamedTuple):
     def to(self, dtype: torch.dtype = None, device: torch.device = None) -> 'LeftAlignedSequence':
         return LeftAlignedSequence(
             data=self.data.to(dtype=dtype, device=device),
-            token_sizes=self.token_sizes.to(dtype=torch.long, device=device),
+            token_sizes=self.token_sizes.to(device=device),
         )
 
     def double(self) -> 'LeftAlignedSequence':
