+++ torchrua/layout/right.py
@@ -14,7 +14,7 @@ class RightAlignedSequence(NamedTuple):
     def to(self, dtype: torch.dtype = None, device: torch.device = None) -> 'RightAlignedSequence':
         return RightAlignedSequence(
             data=self.data.to(dtype=dtype, device=device),
-            token_sizes=self.token_sizes.to(dtype=torch.long, device=device),
+            token_sizes=self.token_sizes.to(device=device),
         )
 
     def double(self) -> 'RightAlignedSequence':
